refactor: 패키지 구조 변경
- 동적 데이터 서빙을 위한 서블릿과 정적 데이터 서빙을 위한 단순 http 서버를 패키지로 구분
