+++ src/main/java/webserver/httpserver/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpserver;
 
 public enum HttpMethod {
     GET, POST
