+++ src/main/java/webserver/httpserver/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
