+++ src/main/java/webserver/httpserver/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.httpserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
